+++ probe-rs/src/architecture/arm/ap/v1/mod.rs
@@ -1,13 +1,11 @@
 //! Types and functions for interacting with the first version of access ports.
-pub(crate) use super::memory as memory_ap;
-
 use crate::architecture::arm::dp::DebugPortError;
 use crate::probe::DebugProbeError;
 
-pub use super::{ApClass, ApType, IDR};
-
-use super::super::{
-    dp::DpAddress, ArmError, DapAccess, FullyQualifiedApAddress, RegisterParseError,
+use super::{memory::{registers::{self, BaseAddrFormat, BASE, BASE2, DRW, TAR, TAR2}, DataSize, MemoryAp}, GenericAp};
+use super::{
+    super::{dp::DpAddress, ArmError, DapAccess, FullyQualifiedApAddress, RegisterParseError},
+    IDR,
 };
 
 /// A trait to be implemented on Access Port register types for typed device access.
@@ -18,6 +16,116 @@ pub trait Register:
     const ADDRESS: u8;
     /// The name of the register as string.
     const NAME: &'static str;
+}
+
+pub(crate) trait MemoryApType:
+    ApRegAccess<BASE> + ApRegAccess<BASE2> + ApRegAccess<TAR> + ApRegAccess<TAR2> + ApRegAccess<DRW>
+{
+    /// This Memory AP’s specific CSW type.
+    type CSW: Register;
+
+    fn has_large_address_extension(&self) -> bool;
+    fn has_large_data_extension(&self) -> bool;
+    fn supports_only_32bit_data_size(&self) -> bool;
+
+    /// Attempts to set the requested data size.
+    ///
+    /// The operation may fail if the requested data size is not supported by the Memory Access
+    /// Port.
+    fn try_set_datasize<I: ApAccess>(
+        &mut self,
+        interface: &mut I,
+        data_size: DataSize,
+    ) -> Result<(), ArmError>;
+
+    /// The current generic CSW (missing the memory AP specific fields).
+    fn generic_status<I: ApAccess>(
+        &mut self,
+        interface: &mut I,
+    ) -> Result<registers::CSW, ArmError> {
+        self.status(interface)?
+            .into()
+            .try_into()
+            .map_err(ArmError::RegisterParse)
+    }
+
+    /// The current CSW with the memory AP specific fields.
+    fn status<I: ApAccess>(&mut self, interface: &mut I) -> Result<Self::CSW, ArmError>;
+
+    /// The base address of this AP which is used to then access all relative control registers.
+    fn base_address<I: ApAccess>(&self, interface: &mut I) -> Result<u64, ArmError> {
+        let base_register: BASE = interface.read_ap_register(self)?;
+
+        let mut base_address = if BaseAddrFormat::ADIv5 == base_register.Format {
+            let base2: BASE2 = interface.read_ap_register(self)?;
+
+            u64::from(base2.BASEADDR) << 32
+        } else {
+            0
+        };
+        base_address |= u64::from(base_register.BASEADDR << 12);
+
+        Ok(base_address)
+    }
+
+    fn set_target_address<I: ApAccess>(
+        &mut self,
+        interface: &mut I,
+        address: u64,
+    ) -> Result<(), ArmError> {
+        let address_lower = address as u32;
+        let address_upper = (address >> 32) as u32;
+
+        if self.has_large_address_extension() {
+            let tar = TAR2 {
+                address: address_upper,
+            };
+            interface.write_ap_register(self, tar)?;
+        } else if address_upper != 0 {
+            return Err(ArmError::OutOfBounds);
+        }
+
+        let tar = TAR {
+            address: address_lower,
+        };
+        interface.write_ap_register(self, tar)?;
+
+        Ok(())
+    }
+
+    /// Read multiple 32 bit values from the DRW register on the given AP.
+    fn read_data<I: ApAccess>(
+        &mut self,
+        interface: &mut I,
+        values: &mut [u32],
+    ) -> Result<(), ArmError> {
+        match values {
+            // If transferring only 1 word, use non-repeated register access, because it might be
+            // faster depending on the probe.
+            [value] => interface.read_ap_register(self).map(|drw: DRW| {
+                *value = drw.data;
+            }),
+            _ => interface.read_ap_register_repeated::<_, DRW>(self, values),
+        }
+        .map_err(AccessPortError::register_read_error::<DRW, _>)
+        .map_err(|err| ArmError::from_access_port(err, self.ap_address()))
+    }
+
+    /// Write multiple 32 bit values to the DRW register on the given AP.
+    fn write_data<I: ApAccess>(
+        &mut self,
+        interface: &mut I,
+        values: &[u32],
+    ) -> Result<(), ArmError> {
+        match values {
+            // If transferring only 1 word, use non-repeated register access, because it might be
+            // faster depending on the probe.
+            &[data] => interface.write_ap_register(self, DRW { data }),
+            _ => interface.write_ap_register_repeated::<_, DRW>(self, values),
+        }
+        .map_err(AccessPortError::register_write_error::<DRW, _>)
+        .map_err(|e| ArmError::from_access_port(e, self.ap_address()))
+    }
 }
 
 /// Some error during AP handling occurred.
@@ -228,7 +336,7 @@ where
 pub enum AccessPort {
     /// Any memory Access Port.
     // TODO: Allow each memory by types to be specialised with there specific feature
-    MemoryAp(memory_ap::MemoryAp),
+    MemoryAp(MemoryAp),
     /// Other Access Ports not used for memory accesses.
     Other(GenericAp),
 }
@@ -292,24 +400,3 @@ where
             }
         })
 }
-
-/// A generic access port which implements just the register every access port has to implement
-/// to be compliant with the ADI 5.2 specification.
-#[derive(Clone, Debug)]
-pub struct GenericAp {
-    address: FullyQualifiedApAddress,
-}
-
-impl GenericAp {
-    /// Creates a new GenericAp with `address` as base address.
-    pub const fn new(address: FullyQualifiedApAddress) -> Self {
-        Self { address }
-    }
-}
-
-impl AccessPortType for GenericAp {
-    fn ap_address(&self) -> &FullyQualifiedApAddress {
-        &self.address
-    }
-}
-impl ApRegAccess<IDR> for GenericAp {}
